requirements.txt
pre-commit
black
ruff
python-dotenv
stripe
litellm
